- update LOG_STATS
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@15074 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/simulation/solver/solver_main.h
@@ -60,11 +60,13 @@ typedef struct SOLVER_INFO
   modelica_boolean didEventStep;
 
   /* stats */
-  unsigned int stateEvents;
-  unsigned int sampleEvents;
-  unsigned int stepNo;
-  unsigned int callsODE;
-  unsigned int callsDAE;
+  unsigned long stateEvents;
+  unsigned long sampleEvents;
+  /* not used
+  unsigned long stepNo;
+  unsigned long callsODE;
+  unsigned long callsDAE;
+  */
 
   void* solverData;
 }SOLVER_INFO;
